Prepare for deployment: final testing and optimizations

+++ src/app/employee/my-leaves/dialogs/delete/delete.component.ts
@@ -2,6 +2,8 @@ import { MAT_DIALOG_DATA, MatDialogRef, MatDialogTitle, MatDialogContent, MatDia
 import { Component, Inject } from '@angular/core';
 import { MyLeavesService } from '../../my-leaves.service';
 import { MatButtonModule } from '@angular/material/button';
+import { Status } from '../../my-leaves.model';
+import Swal from 'sweetalert2';
 
 export interface DialogData {
   _id: string;
@@ -32,10 +34,34 @@ export class DeleteDialogComponent {
   onNoClick(): void {
     this.dialogRef.close();
   }
-  confirmDelete(): void {
-    const leave = this.data._id ; // Convertir _id en chaîne de caractères
+  // confirmDelete(): void {
+  //   const leave = this.data._id ; // Convertir _id en chaîne de caractères
+  //   this.myLeavesService.deleteMyLeaves(leave);
+  // }
+
+confirmDelete(): void {
+    const leave = this.data._id; // Convertir _id en chaîne de caractères
+    const status = this.data.status;
+
+    if (status === Status.Approved || status === Status.Declined) {
+        // Afficher une alerte SweetAlert pour informer l'utilisateur que le leave ne peut pas être supprimé
+        Swal.fire({
+            icon: 'error',
+            title: 'Unable to delete',
+            text: 'Cannot delete , This leave has been processed by the administrator.',
+            confirmButtonText: 'OK'
+        });
+        return;
+    }
+    // Supprimer le leave s'il n'est pas approuvé ou refusé
     this.myLeavesService.deleteMyLeaves(leave);
-  }
-  
+    Swal.fire({
+      icon: 'success',
+      title: 'Leave deleted',
+      text: 'The leave has been successfully deleted.',
+      confirmButtonText: 'OK'
+  });
+}
+
   
 }
